[DEV-285] feat: 여러 용어 한번에 히스토리에 추가할 수 있는 API 추가
- 요청으로 들어온 용어 들의 히스토리 유무 확인 후 히스토리 추가

+++ src/main/java/com/onedreamus/project/thisismoney/model/dto/history/DictionaryIdDto.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @Setter
 public class DictionaryIdDto {
 
-    private Integer id; // 용어 ID
+    private Long id; // 용어 ID
 }
